backend/requirements.txt
# Unified UsenetSync Requirements

# Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Database
psycopg2-binary==2.9.9
redis==5.0.1
alembic==1.12.1

# Security
cryptography==41.0.7
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# Networking
httpx==0.25.2
websockets==12.0
aiofiles==23.2.1

# Usenet/NNTP (CRITICAL!)
pynntp==1.0.2

# Monitoring
psutil==5.9.6
prometheus-client==0.19.0

# Utilities
python-dotenv==1.0.0
click==8.1.7
rich==13.7.0
pyyaml==6.0.1

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
mypy==1.7.1
ruff==0.1.6